Move ExportDatabaseCSVTask into a new DatabaseExporter file
